Add 'nflatten', the narwhal flatten tool
The existing flatten tool is an objj script that relies on the
interesting rhino ability to instantiate another javascript vm within a
process that is already a javascript vm.

The C++ interface to V8 would let one do the same thing but we don't use
that technology directly in the cappuccino project so a more straight
forward approach is taken.

nflatten is a copy of flatten, but modified to run and work as a narwhal
script.  The new objj-analysis-tools.js file is a copy of the .j form.
Where the .j form put the browser objective-j environment into the rhino
sub-javascript vm, this loads the objective-j environment into the
CommonJS narwhal environment, with enough browser-like hooks to make the
parser and compiler work as if they were running in the browser
environment.

The original files are unchanged so one can still use the rhino version
of flatten.

The original author of the flatten script should get the credit in git
for most of the lines in these new files but I don't know how to
accomplish that while maintaining the original files.